txt handler ver 0.2
Reading from txt file reimplemented according to the last consultation.

The TxtHandler.SEPARATOR class variable definies the separation between
two paragraphs. If the separation number is even - for example: N -, in
the output file it will appears as N+1 separation, because every "first"
separation line is handled as separation and the "second" is as a
paragraph.